Fix multiple file write bug
Simply moving the vcb_update_free() call up fixes this bug because
originally we were writing the new inode into the free block, and then
trying to update the vcb by reading the formerly free block as a free
block, which did not work. To fix this, we update the vcb while the
block is still free, then do the write.